integrated front and backend, added new file structure
